Format the number of supported param conversions as floats
This makes no sense, but it may make it a bit more obvious what's going
on if this is not an integer. Airwin2clap currently for instance reports
0.33333334 or 0.6666667 on this metric for most of its parameters.

+++ src/tests/plugin.rs
@@ -401,11 +401,14 @@ impl<'a> TestCase<'a> for PluginTestCase {
                         }
 
                         if num_supported_value_to_text != 0 && num_supported_value_to_text != expected_conversions {
-                            let num_supported_params = num_supported_value_to_text / VALUES_PER_PARAM;
+                            // This makes zero sense as a float, but when plugins are doing some
+                            // _really_ weird things getting a value like 0.33334 instead of 0 may
+                            // make it a bit more obvious what's goin on
+                            let num_supported_params = num_supported_value_to_text as f32 / VALUES_PER_PARAM as f32;
                             anyhow::bail!("'clap_plugin_params::value_to_text()' is supported for {num_supported_params} out of {num_params} parameters. This function is expected to be supported for either none of the parameters or for all of them.");
                         }
                         if num_supported_text_to_value != 0 && num_supported_text_to_value != expected_conversions {
-                            let num_supported_params = num_supported_text_to_value / VALUES_PER_PARAM;
+                            let num_supported_params = num_supported_text_to_value as f32 / VALUES_PER_PARAM as f32;
                             anyhow::bail!("'clap_plugin_params::text_to_value()' is supported for {num_supported_params} out of {num_params} parameters. This function is expected to be supported for either none of the parameters or for all of them.");
                         }
